requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.26.1
azure-functions==1.12.0
azure-storage-blob==12.14.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz       
et-xmlfile==1.1.0
fonttools==4.38.0
fst-pso==1.8.1
funcy==1.17
future==0.18.2
FuzzyTM==2.0.5
gensim==4.3.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
miniful==0.0.6
msrest==0.7.1
murmurhash==1.0.9
nltk==3.8
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.9
packaging==22.0
pandas==1.5.2
pathy==0.10.1
Pillow==9.3.0
preshed==3.0.8
pycparser==2.21
pydantic==1.8.2
pyFUME==0.2.25
pyLDAvis==3.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.2.0
scipy==1.9.3
simpful==2.9.0
six==1.16.0
sklearn==0.0
smart-open==6.3.0
spacy==3.0.9
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
textblob==0.17.1
thinc==8.0.17
threadpoolctl==3.1.0
tqdm==4.64.1
typer==0.3.2
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
wordcloud==1.8.2.2